+++ plugins/logic/src/notification_entity.cpp
@@ -1,4 +1,7 @@
 #include "notification_entity.h"
 
-LogicPlugin::NotificationEntity::NotificationEntity(QString title, QString description, int type): AbstractNotification(title,  description, type)
+LogicPlugin::NotificationEntity::NotificationEntity(QString title,
+                                                    QString description,
+                                                    uint16_t type) :
+    AbstractNotification(title, description, type)
 {}
